requirements.txt
tensorflow-addons==0.15.0
tensorflow-gpu==2.2.0
transformers==4.9.1
tqdm==4.31.1
pandas==1.3.5
scikit-learn==1.0.2